on adding an item to cart, update if item exists on table else create new record. added success message on item page
