+++ public/assets/css/style-item.css
@@ -95,3 +95,8 @@
   text-decoration: underline;
   font-size: 1.5rem;
 }
+
+.addToCartMessage {
+  color: var(--primary-solid);
+  margin-top: 1rem;
+}
