+++ src/views/index.ejs
@@ -54,6 +54,6 @@
     <script src="https://cdn.jsdelivr.net/npm/iconify-icon@1.0.8/dist/iconify-icon.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@glidejs/glide"></script>
     <script src="/assets/js/slider.js"></script>
-    <script src="/assets/js/cart.js"></script>
+    <script src="/assets/js/item.js"></script>
   </body>
 </html>
