+++ src/views/shop/cart.ejs
@@ -103,6 +103,6 @@
     <script src="/assets/js/slider.js"></script>
     <script src="/assets/js/buttons.js"></script>
     <script src="/assets/js/calculateTotal.js"></script>
-    <script src="/assets/js/cart.js"></script>
+    <script src="/assets/js/item.js"></script>
   </body>
 </html>
