+++ src/views/shop/item.ejs
@@ -56,6 +56,7 @@
               <a href="#">Ver métodos de pago</a> - <%= product.dues %> CUOTAS
               SIN INTERES
             </p>
+            <p class="addToCartMessage"></p>
           </article>
         </div>
       </form>
@@ -69,6 +70,6 @@
     <script src="https://cdn.jsdelivr.net/npm/@glidejs/glide"></script>
     <script src="/assets/js/buttons.js"></script>
     <script src="/assets/js/slider.js"></script>
-    <script src="/assets/js/cart.js"></script>
+    <script src="/assets/js/item.js"></script>
   </body>
 </html>
